backend/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["caio pilon <caiofpilon@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.79"
uvicorn = "^0.18"
SQLAlchemy = "^1.4"
psycopg2-binary = "^2.9"
python-dotenv = "^0.20"
celery = "^5.1"
flower = "^1.0"
redis = "^3.5"

[tool.poetry.dev-dependencies]
coverage = "^6.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
